pyproject.toml
[project]
name = "projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython>=9.9.0",
    "langsmith>=0.6.2",
    "openai>=2.14.0",
    "openai-agents>=0.6.5",
    "python-dotenv>=1.2.1",
    "tavily>=1.1.0",
]
